backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# FREE AI Services
google-generativeai==0.3.2
langchain==0.1.6
langchain-google-genai==0.0.6

# FREE Vector Database
chromadb==0.4.22

# Embeddings (Runs locally - FREE)
sentence-transformers==2.3.1

# Database (FREE 500MB)
supabase==2.3.4

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# CORS
starlette==0.35.1
